+++ Event/AbstractUserEvent.php
@@ -33,16 +33,4 @@ abstract class AbstractUserEvent extends Event
     {
         return $this->user;
     }
-
-    /**
-     * Checks whether a user is available.
-     *
-     * @return bool
-     *
-     * @deprecated should be used in the `OnInvalidCredentialsEvent` only, will be removed from here in 2.0
-     */
-    public function isUserAvailable()
-    {
-        return null !== $this->user;
-    }
 }